+++ core/context_processors.py
@@ -37,11 +37,15 @@ def services_home_links(request):
             'url': reverse_lazy('invest-home'),
             'label': _('Invest')
         },
+        'investment_atlas_home_link': {
+            'url': reverse_lazy('atlas-home'),
+            'label': _('Invest in the UK')
+        },
     }
 
 
 def header_navigation(request):
-    nav_tree = header_config.nav_tree.HEADER_TREE \
+    nav_tree = header_config.nav_tree.ATLAS_HEADER_TREE \
         if settings.FEATURE_FLAGS['NEW_IA_ON'] \
         else header_config.nav_tree.OLD_HEADER_TREE